requirements.txt
PyQt6>=6.5
pydub>=0.25.1
numpy>=1.24
openai>=1.0
requests>=2.31
pyannote.audio>=3.1
torch>=2.0
torchaudio>=2.0
openai-whisper>=20231117
python-docx>=1.0
Pillow>=10.0
pyinstaller>=6.0
